+++ pages/LoginPage/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 // import './common.css';
-import "./LoginPage.css";
+import "../../styles/LoginPage.css";
 
 // 이미지 import
 import Logo from "../../public/images/logo.svg";
