+++ pages/boards/index.tsx
@@ -1,0 +1,1 @@
+
