+++ styles/common.css
@@ -1,4 +1,4 @@
-* {
+/* * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
@@ -20,4 +20,4 @@ button {
 img,
 svg {
   vertical-align: bottom;
-}
+} */
